move client response to own module
